Ready to update and write index
Restructured task_save_data_t and the checks for multi-frame streams.
We now account for MCAs as well as traces, checking for extra or missed
frames, or interrupts by other frames (e.g. non-trace in the middle of
a trace). TO DO: check that all trace frames in a stream are the same
type.