Added a library to convert captures to hdf5 files
Eventually it will be used by the save-to-file task. The request to save
all frames to file will then include two additional fields: measurement
and wait. After capture is complete, all streams (including indices) are
saved in a single hdf5 file as separate datasets under a group
<measurement>. If wait is false, the task will reply to the client when
conversion beings, otherwise---when it ends.