Removed 'noexit' version as it's unsafe
If after successful second fork, caller signals with DAEMON_ERR_MSG but
returns, rather than exit, code after call to daemonize_noexit will be
run twice. Intent was a functionality to call a custom function before
parent exits. This will be provided by accepting a pointer to a handler
and an argument soon.