Changed the request to hdf5_conv
No longer using offset to decide whether to mmap file. Filename and
buffer are no longer a union. Exactly one of filename and buffer must be
set, so we no longer clobber the pointer (commit
b75aefeab46a9cef4a723d4d3b47fe3af6cf5f6b didn't achieve that,
obviously...). Also, dropped the const qualifier, not necessary. The
only part of the given request that can change are the length and
offset for datasets which are not mmapped.

1) If filename is NULL, copy directly from buffer.
   Length and offset must then be non-negative.

2) If buffer is NULL, open filename and mmap it. Will close the file and
   unmap at the end.
   If offset < 0 it is takes with respect to EOF.
   If length < 0 or extends beyond EOF, copy until EOF.
   If length == 0 or offset extends beyond EOF, the dataset will be empty.
   The value of offset and length after the call is unspecified.