Added the client option to relink group only
A new option, '-r' for the 'save-to-file' task in tesc will request
the measurement group to be moved. Conflicts with '-o' which requests
the file to be overwritten.

Also moved the check for the value of ovrwrtmode in the client request
from tesd_task_save (as soon as the request comes) to hdf5conv (when
conversion begins).