Added a request for conversion
A new frame requests conversion of a previous job.

Also, all files are unlinked instead of overwritten (unless overwriting
is forbidden by the request, in which case nothing is written anyway).
This will prevent strange errors if the files are open in another
process, prevent writing outside of data root if any of the index or
data files are symlinks, and prevent permission errors if the file was
created as another user (as long as the current user has write
permission to the directory, which he should).