Fixed wrong order of header inclusion
tespkt.h checks for TES_MCASIZE_BUG which was defined in api.h, which
was included after tespkt.h... This flag was only needed to define the
maximum size of a histogram. The flag, max size of a histogram and
average trace are now defined in tespkt.h as they should be.