Bug fixes in tick publishing
Ticks that arrive after the last event in a coincidence are now
published without the UNRESOLVED flag, even if they were within the
coincidence window. I.e. if a ch1, ch0, tick comes and the tick ends the
coincidence, it's not unresolved.
Also if multiple ticks come during a coincidence window, all that are
before the last event are UNRESOLVED, all others are resolved.

Also fixed: ensuring there's enough room in the buffer for the ticks.