Bug fixes
canonicalize_path no longer assumes root exists and is not a symlink.
It now correctly handles the case of symlinked or non-existing root
(but not dangling symlinks, that's the intended behaviour).

Moved bulk of it to mkdirr which now ignores trailing slash and creates
the final component eitherway.

tesd.c ensures there's a slash at end of confdir so it's created by
canonicalize_path (which would otherwise call mkdirr on the directory
part only)