Subscription zlistx_t -> zhashx_t
It doesn't make sense (and is not doable) to have the task process the
subscription pattern and return something else to be save in the list,
since the item comparator is called when we lookup a pattern string and
also when reordering the list (implicitly or explicitly). In the former
case the second argument would be a string, in the latter would be
whatever type the validator returns.

A hash table is what is applicable here.

Also fixed: endp_(un)subscribe need to first call s_endp_sub_del/add
and only then (if not deadbeef) try to update the socket state.

Also added the processing of the pattern string in tesd_task_coinccount